The whole code refactored and only one user entity used
