+++ src/app.module.ts
@@ -7,11 +7,6 @@ import { AuthModule } from './modules/auth/auth.module';
 import { UsersModule } from './modules/users/users.module';
 import { SmsModule } from './shared/sms/sms.module';
 import { UploadModule } from './shared/upload/upload.module';
-import { Employer } from './modules/users/entities/employer.entity';
-import { ServiceProvider } from './modules/users/entities/serviceProvider.entity';
-import { PropertyOwner } from './modules/users/entities/propertyOwner.entity';
-import { PropertyRenter } from './modules/users/entities/propertyRenter.entity';
-import { BabySitterFinder } from './modules/users/entities/babySitterFinder.entity';
 import { RefreshToken } from './modules/auth/entities/refresh-token.entity';
 import { User } from './modules/users/entities/users.entity';
 
@@ -30,15 +25,7 @@ import { User } from './modules/users/entities/users.entity';
         username: config.get<string>('DATABASE_USERNAME'),
         password: config.get<string>('DATABASE_PASSWORD'),
         database: config.get<string>('DATABASE_NAME'),
-        entities: [
-          User,
-          Employer,
-          ServiceProvider,
-          PropertyOwner,
-          PropertyRenter,
-          BabySitterFinder,
-          RefreshToken,
-        ],
+        entities: [User, RefreshToken],
         synchronize: true,
         ssl:
           config.get<string>('NODE_ENV') !== 'development'
