+++ src/modules/auth/auth.service.ts
@@ -14,27 +14,15 @@ import { CustomMailerService } from 'src/shared/mailer/mailer.service';
 import { accountVerificationEmail } from 'src/shared/mailer/templates/account-verification.template';
 import { SmsService } from 'src/shared/sms/sms.service';
 import { UploadService } from 'src/shared/upload/upload.service';
-import { Employer } from '../users/entities/employer.entity';
-import { ServiceProvider } from '../users/entities/serviceProvider.entity';
-import { PropertyOwner } from '../users/entities/propertyOwner.entity';
-import { PropertyRenter } from '../users/entities/propertyRenter.entity';
 import { RefreshToken } from './entities/refresh-token.entity';
 import { capitalizeString } from 'src/shared/utils/capitilize-string.util';
-import { BabySitterFinder } from '../users/entities/babySitterFinder.entity';
+import { User } from '../users/entities/users.entity';
 
 @Injectable()
 export class AuthService {
   constructor(
-    @InjectRepository(Employer)
-    private readonly employerRepo: Repository<Employer>,
-    @InjectRepository(ServiceProvider)
-    private readonly serviceProviderRepo: Repository<ServiceProvider>,
-    @InjectRepository(PropertyOwner)
-    private readonly propertyOwnerRepo: Repository<PropertyOwner>,
-    @InjectRepository(PropertyRenter)
-    private readonly propertyRenterRepo: Repository<PropertyRenter>,
-    @InjectRepository(BabySitterFinder)
-    private readonly babySitterFinderRepo: Repository<BabySitterFinder>,
+    @InjectRepository(User)
+    private readonly userRepo: Repository<User>,
     @InjectRepository(RefreshToken)
     private readonly refreshTokenRepo: Repository<RefreshToken>,
     private readonly jwtService: JwtService,
@@ -178,13 +166,7 @@ export class AuthService {
   }
 
   async findUserTypeById(id: string): Promise<string> {
-    const user =
-      (await this.employerRepo.findOne({ where: { id } })) ||
-      (await this.serviceProviderRepo.findOne({ where: { id } })) ||
-      (await this.propertyOwnerRepo.findOne({ where: { id } })) ||
-      (await this.propertyRenterRepo.findOne({ where: { id } })) ||
-      (await this.babySitterFinderRepo.findOne({ where: { id } }));
-
+    const user = await this.userRepo.findOne({ where: { id } });
     if (!user) {
       throw new NotFoundException('User not found');
     }
@@ -193,29 +175,7 @@ export class AuthService {
   }
 
   async findUserByEmailOrPhone(email: string, phoneNumber: string) {
-    let user: any;
-
-    user = await this.employerRepo.findOne({
-      where: [{ email }, { phoneNumber }],
-    });
-    if (user) return user;
-
-    user = await this.serviceProviderRepo.findOne({
-      where: [{ email }, { phoneNumber }],
-    });
-    if (user) return user;
-
-    user = await this.propertyOwnerRepo.findOne({
-      where: [{ email }, { phoneNumber }],
-    });
-    if (user) return user;
-
-    user = await this.propertyRenterRepo.findOne({
-      where: [{ email }, { phoneNumber }],
-    });
-    if (user) return user;
-
-    user = await this.babySitterFinderRepo.findOne({
+    const user = await this.userRepo.findOne({
       where: [{ email }, { phoneNumber }],
     });
     if (user) return user;
@@ -227,105 +187,26 @@ export class AuthService {
 
   // ⁡⁢⁢⁡⁢⁢⁢𝗜𝗡𝗜𝗧𝗜𝗔𝗟 𝗦𝗜𝗚𝗡𝗨𝗣⁡
   async signup(body: any) {
-    if (body.userType === 'employer') {
-      const hashedPassword = await this.hashPassword(
-        body.password,
-        body.confirmPassword,
-      );
-      // Create the user and save to the database
-      const user = this.employerRepo.create({
-        userType: body.userType,
-        email: body.email,
-        phoneNumber: body.phoneNumber,
-        password: hashedPassword,
-      });
-      await this.employerRepo.save(user);
-      await this.generateAndSendVerificationCode(this.employerRepo, user.id);
-      return user;
-    } else if (body.userType === 'serviceProvider') {
-      const hashedPassword = await this.hashPassword(
-        body.password,
-        body.confirmPassword,
-      );
-      // Create the user and save to the database
-      const user = this.serviceProviderRepo.create({
-        userType: body.userType,
-        email: body.email,
-        phoneNumber: body.phoneNumber,
-        password: hashedPassword,
-      });
-      await this.serviceProviderRepo.save(user);
-      await this.generateAndSendVerificationCode(
-        this.serviceProviderRepo,
-        user.id,
-      );
-      return user;
-    } else if (body.userType === 'propertyOwner') {
-      const hashedPassword = await this.hashPassword(
-        body.password,
-        body.confirmPassword,
-      );
-      // Create the user and save to the database
-      const user = this.propertyOwnerRepo.create({
-        userType: body.userType,
-        email: body.email,
-        phoneNumber: body.phoneNumber,
-        password: hashedPassword,
-      });
-      await this.propertyOwnerRepo.save(user);
-      await this.generateAndSendVerificationCode(
-        this.propertyOwnerRepo,
-        user.id,
-      );
-      return user;
-    } else if (body.userType === 'propertyRenter') {
-      const hashedPassword = await this.hashPassword(
-        body.password,
-        body.confirmPassword,
-      );
-      // Create the user and save to the database
-      const user = this.propertyRenterRepo.create({
-        userType: body.userType,
-        email: body.email,
-        phoneNumber: body.phoneNumber,
-        password: hashedPassword,
-      });
-      await this.propertyRenterRepo.save(user);
-      await this.generateAndSendVerificationCode(
-        this.propertyRenterRepo,
-        user.id,
-      );
-      return user;
-    } else if (body.userType === 'babySitterFinder') {
-      const hashedPassword = await this.hashPassword(
-        body.password,
-        body.confirmPassword,
-      );
-      // Create the user and save to the database
-      const user = this.babySitterFinderRepo.create({
-        userType: body.userType,
-        email: body.email,
-        phoneNumber: body.phoneNumber,
-        password: hashedPassword,
-      });
-      await this.babySitterFinderRepo.save(user);
-      await this.generateAndSendVerificationCode(
-        this.babySitterFinderRepo,
-        user.id,
-      );
-      return user;
-    } else {
-      return false;
-    }
+    const hashedPassword = await this.hashPassword(
+      body.password,
+      body.confirmPassword,
+    );
+    // Create the user and save to the database
+    const user = this.userRepo.create({
+      userType: body.userType,
+      email: body.email,
+      phoneNumber: body.phoneNumber,
+      password: hashedPassword,
+    });
+    await this.userRepo.save(user);
+    await this.generateAndSendVerificationCode(user.id);
+    return user;
   }
 
   // ⁡⁢⁣⁣⁡⁢⁢⁢⁡⁢⁢⁢𝗔𝗖𝗖𝗢𝗨𝗡𝗧 𝗩𝗘𝗥𝗜𝗙𝗜𝗖𝗔𝗧𝗜𝗢𝗡⁡
   // Send 6-digit SMS or email verification code to the user
-  async generateAndSendVerificationCode(
-    repo: Repository<any>,
-    userId: string,
-  ): Promise<void> {
-    const user = await repo.findOne({
+  async generateAndSendVerificationCode(userId: string): Promise<void> {
+    const user = await this.userRepo.findOne({
       where: { id: userId },
     });
 
@@ -344,7 +225,7 @@ export class AuthService {
       Date.now() + expiryTime * 60 * 1000,
     );
 
-    await repo.update(userId, {
+    await this.userRepo.update(userId, {
       verificationCode,
       verificationCodeExpires,
     });
@@ -379,8 +260,8 @@ export class AuthService {
     }
   }
 
-  async verifyAccount(repo: Repository<any>, userId: string, code: string) {
-    const user = await repo.findOne({ where: { id: userId } });
+  async verifyAccount(userId: string, code: string) {
+    const user = await this.userRepo.findOne({ where: { id: userId } });
     if (!user) {
       throw new NotFoundException('Employer not found');
     }
@@ -397,7 +278,7 @@ export class AuthService {
       throw new BadRequestException('Verification code has expired');
     }
 
-    await repo.update(userId, {
+    await this.userRepo.update(userId, {
       isVerified: true,
       verificationCode: null,
       verificationCodeExpires: null,
@@ -415,13 +296,12 @@ export class AuthService {
 
   // ⁡⁢⁣⁣⁡⁢⁢⁢𝗣𝗥𝗢𝗙𝗜𝗟𝗘 𝗖𝗢𝗠𝗣𝗟𝗘𝗧𝗜𝗢𝗡 𝗙𝗢𝗥 ⁡⁢⁢⁢𝗘𝗠𝗣𝗟𝗢𝗬𝗘𝗥⁡
   async completeEmployerProfile(
-    repo: Repository<any>,
     userId: string,
     body: any,
     profileImg: Express.Multer.File,
   ) {
     // Check if the user exists
-    const user = await repo.findOne({ where: { id: userId } });
+    const user = await this.userRepo.findOne({ where: { id: userId } });
     if (!user) {
       throw new NotFoundException('User not found');
     }
@@ -438,7 +318,7 @@ export class AuthService {
     }
 
     // Update the fields
-    const updatedUser = await repo.update(userId, {
+    const updatedUser = await this.userRepo.update(userId, {
       firstName: capitalizeString(body.firstName),
       lastName: capitalizeString(body.lastName),
       email: user.email ? user.email : body.email,
@@ -450,7 +330,7 @@ export class AuthService {
       bio: body.bio,
     });
 
-    await repo.update(userId, { isProfileCompleted: true });
+    await this.userRepo.update(userId, { isProfileCompleted: true });
 
     // ⁡⁢⁢⁢⁡⁢⁢⁢𝗦𝗨𝗕𝗦𝗖𝗥𝗜𝗣𝗧𝗜𝗢𝗡⁡
 
@@ -461,14 +341,13 @@ export class AuthService {
 
   // ⁡⁢⁣⁣⁡⁢⁢⁢⁡⁢⁢⁢𝗣𝗥𝗢𝗙𝗜𝗟𝗘 𝗖𝗢𝗠𝗣𝗟𝗘𝗧𝗜𝗢𝗡 𝗙𝗢𝗥 𝗦𝗘𝗥𝗩𝗜𝗖𝗘 𝗣𝗥𝗢𝗩𝗜𝗗𝗘𝗥𝗦⁡
   async completeServiceProvidersProfile(
-    repo: Repository<any>,
     userId: string,
     body: any,
     profileImg: Express.Multer.File,
     portfolioFiles: Express.Multer.File[],
   ) {
     // Check if the user exists
-    const user = await repo.findOne({ where: { id: userId } });
+    const user = await this.userRepo.findOne({ where: { id: userId } });
     if (!user) {
       throw new NotFoundException('User not found');
     }
@@ -500,7 +379,7 @@ export class AuthService {
     }
 
     // Update the fields
-    const updatedUser = await repo.update(userId, {
+    const updatedUser = await this.userRepo.update(userId, {
       firstName: capitalizeString(body.firstName),
       lastName: capitalizeString(body.lastName),
       email: user.email ? user.email : body.email,
@@ -520,7 +399,7 @@ export class AuthService {
       hourlyRate: body.hourlyRate,
     });
 
-    await repo.update(userId, { isProfileCompleted: true });
+    await this.userRepo.update(userId, { isProfileCompleted: true });
 
     // ⁡⁢⁢⁢⁡⁢⁢⁢𝗦𝗨𝗕𝗦𝗖𝗥𝗜𝗣𝗧𝗜𝗢𝗡⁡
 
@@ -531,13 +410,12 @@ export class AuthService {
 
   // ⁡⁢⁣⁣⁡⁢⁢⁢⁡⁢⁢⁢𝗣𝗥𝗢𝗙𝗜𝗟𝗘 𝗖𝗢𝗠𝗣𝗟𝗘𝗧𝗜𝗢𝗡 𝗙𝗢𝗥 𝗣𝗥𝗢𝗣𝗘𝗥𝗧𝗬 𝗢𝗪𝗡𝗘𝗥𝗦⁡
   async completePropertyOwnersProfile(
-    repo: Repository<any>,
     userId: string,
     body: any,
     profileImg: Express.Multer.File,
   ) {
     // Check if the user exists
-    const user = await repo.findOne({ where: { id: userId } });
+    const user = await this.userRepo.findOne({ where: { id: userId } });
     if (!user) {
       throw new NotFoundException('User not found');
     }
@@ -554,7 +432,7 @@ export class AuthService {
     }
 
     // Update the fields
-    const updatedUser = await repo.update(userId, {
+    const updatedUser = await this.userRepo.update(userId, {
       firstName: capitalizeString(body.firstName),
       lastName: capitalizeString(body.lastName),
       email: user.email ? user.email : body.email,
@@ -566,7 +444,7 @@ export class AuthService {
       propertyType: body.propertyType,
     });
 
-    await repo.update(userId, { isProfileCompleted: true });
+    await this.userRepo.update(userId, { isProfileCompleted: true });
 
     // ⁡⁢⁢⁢𝗦𝗨𝗕𝗦𝗖𝗥𝗜𝗣𝗧𝗜𝗢𝗡⁡
 
@@ -577,13 +455,12 @@ export class AuthService {
 
   // ⁡⁢⁢⁢𝗣𝗥𝗢𝗙𝗜𝗟𝗘 𝗖𝗢𝗠𝗣𝗟𝗘𝗧𝗜𝗢𝗡 𝗙𝗢𝗥 𝗣𝗥𝗢𝗣𝗘𝗥𝗧𝗬 𝗥𝗘𝗡𝗧𝗘𝗥⁡
   async completePropertyRenterProfile(
-    repo: Repository<any>,
     userId: string,
     body: any,
     profileImg: Express.Multer.File,
   ) {
     // Check if the user exists
-    const user = await repo.findOne({ where: { id: userId } });
+    const user = await this.userRepo.findOne({ where: { id: userId } });
     if (!user) {
       throw new NotFoundException('User not found');
     }
@@ -600,7 +477,7 @@ export class AuthService {
     }
 
     // Update the fields
-    const updatedUser = await repo.update(userId, {
+    const updatedUser = await this.userRepo.update(userId, {
       firstName: capitalizeString(body.firstName),
       lastName: capitalizeString(body.lastName),
       email: user.email ? user.email : body.email,
@@ -612,7 +489,7 @@ export class AuthService {
       budgetRange: body.budgetRange,
     });
 
-    await repo.update(userId, { isProfileCompleted: true });
+    await this.userRepo.update(userId, { isProfileCompleted: true });
 
     // ⁡⁢⁢⁢𝗦𝗨𝗕𝗦𝗖𝗥𝗜𝗣𝗧𝗜𝗢𝗡⁡
 
@@ -623,13 +500,12 @@ export class AuthService {
 
   // ⁡⁢⁢⁢⁡⁢⁢⁢𝗣𝗥𝗢𝗙𝗜𝗟𝗘 𝗖𝗢𝗠𝗣𝗟𝗘𝗧𝗜𝗢𝗡 𝗙𝗢𝗥 𝗕𝗔𝗕𝗬 𝗦𝗜𝗧𝗧𝗘𝗥 𝗙𝗜𝗡𝗗𝗘𝗥⁡
   async completeBabySitterFinderProfile(
-    repo: Repository<any>,
     userId: string,
     body: any,
     profileImg: Express.Multer.File,
   ) {
     // Check if the user exists
-    const user = await repo.findOne({ where: { id: userId } });
+    const user = await this.userRepo.findOne({ where: { id: userId } });
     if (!user) {
       throw new NotFoundException('User not found');
     }
@@ -646,7 +522,7 @@ export class AuthService {
     }
 
     // Update the fields
-    const updatedUser = await repo.update(userId, {
+    const updatedUser = await this.userRepo.update(userId, {
       firstName: capitalizeString(body.firstName),
       lastName: capitalizeString(body.lastName),
       email: user.email ? user.email : body.email,
@@ -657,7 +533,7 @@ export class AuthService {
       location: body.location,
     });
 
-    await repo.update(userId, { isProfileCompleted: true });
+    await this.userRepo.update(userId, { isProfileCompleted: true });
 
     // ⁡⁢⁢⁢𝗦𝗨𝗕𝗦𝗖𝗥𝗜𝗣𝗧𝗜𝗢𝗡⁡
