--- src/modules/users/entities/employer.entity.ts
@@ -1,24 +1,0 @@
-import {
-  Entity,
-  Column,
-  CreateDateColumn,
-  UpdateDateColumn,
-  OneToMany,
-} from 'typeorm';
-import { User } from './users.entity';
-import { RefreshToken } from 'src/modules/auth/entities/refresh-token.entity';
-
-@Entity()
-export class Employer extends User {
-  @Column({ nullable: true })
-  companyName: string;
-
-  @CreateDateColumn()
-  createdAt: Date;
-
-  @UpdateDateColumn()
-  updatedAt: Date;
-
-  @OneToMany(() => RefreshToken, (refreshToken) => refreshToken.user)
-  refreshTokens: RefreshToken[];
-}
