--- src/modules/users/entities/propertyRenter.entity.ts
@@ -1,27 +1,0 @@
-import {
-  Entity,
-  Column,
-  CreateDateColumn,
-  UpdateDateColumn,
-  OneToMany,
-} from 'typeorm';
-import { User } from './users.entity';
-import { RefreshToken } from 'src/modules/auth/entities/refresh-token.entity';
-
-@Entity()
-export class PropertyRenter extends User {
-  @Column({ type: 'jsonb', nullable: true })
-  budgetRange: {
-    min: number;
-    max: number;
-  };
-
-  @CreateDateColumn()
-  createdAt: Date;
-
-  @UpdateDateColumn()
-  updatedAt: Date;
-
-  @OneToMany(() => RefreshToken, (refreshToken) => refreshToken.user)
-  refreshTokens: RefreshToken[];
-}
