+++ src/modules/users/entities/users.entity.ts
@@ -1,9 +1,28 @@
+import { RefreshToken } from 'src/modules/auth/entities/refresh-token.entity';
 import {
   Column,
+  CreateDateColumn,
   Entity,
+  OneToMany,
   PrimaryGeneratedColumn,
   TableInheritance,
+  UpdateDateColumn,
 } from 'typeorm';
+
+interface Qualifications {
+  degree: string;
+  certifications: string[];
+  trainings: string[];
+}
+
+interface Experience {
+  positionn: string;
+  yearsOfExperience: string;
+  responsibilities: string[];
+  company?: string;
+  startDate?: Date;
+  endDate?: Date;
+}
 
 @Entity('users')
 @TableInheritance({ column: { type: 'varchar' } })
@@ -57,15 +76,75 @@ export abstract class User {
   @Column({ type: 'text', nullable: true })
   bio: string;
 
-  @Column({ default: 'unsubscribed' })
-  subscriptionStatus: 'subscribed' | 'unsubscribed';
+  @Column({ nullable: true })
+  companyName: string;
 
-  @Column({ default: 'no-plan' })
-  subscriptionPlan: 'per-post | monthly' | 'six-month' | 'Yearly';
+  @Column({ nullable: true })
+  profession: string;
+
+  @Column({ nullable: true })
+  serviceCategory: string;
+
+  @Column({ nullable: true })
+  serviceTitle: string;
+
+  @Column({ type: 'jsonb', nullable: true })
+  experience: Experience[];
+
+  @Column({ type: 'jsonb', nullable: true })
+  qualifications: Qualifications;
+
+  @Column({ type: 'jsonb', nullable: true })
+  skills: string[];
+
+  @Column({ type: 'jsonb', nullable: true })
+  portfolioLinks: string[];
+
+  @Column({ type: 'jsonb', nullable: true })
+  portfolioFiles: string[];
+
+  @Column({ type: 'decimal', precision: 5, scale: 2, nullable: true })
+  hourlyRate: number;
+
+  @Column({ type: 'jsonb', nullable: true })
+  verificationDocuments: string[];
+
+  @Column({ type: 'jsonb', nullable: true })
+  availability: {
+    days: string[];
+    hours: string;
+  };
+
+  @Column({ type: 'jsonb', nullable: true })
+  languages: string[];
+
+  @Column({ nullable: true })
+  propertyType: string;
+
+  @Column({ type: 'jsonb', nullable: true })
+  budgetRange: {
+    min: number;
+    max: number;
+  };
 
   @Column({ type: 'int', nullable: true })
   ratings: number;
 
   @Column({ type: 'jsonb', nullable: true })
   reviews: string[];
+
+  @Column({ default: 'unsubscribed' })
+  subscriptionStatus: 'subscribed' | 'unsubscribed';
+
+  @Column({ default: 'no-plan' })
+  subscriptionPlan: 'per-post | monthly' | 'six-month' | 'Yearly';
+
+  @CreateDateColumn()
+  createdAt: Date;
+
+  @UpdateDateColumn()
+  updatedAt: Date;
+
+  @OneToMany(() => RefreshToken, (refreshToken) => refreshToken.user)
+  refreshTokens: RefreshToken[];
 }
